requirements.txt
fastapi[standard] ==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.46
python-multipart==0.0.6
pytest==7.4.3
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
starlette==0.27.0
python-dotenv==1.0.0
pydantic-settings==2.13.0
bcrypt==5.0.0
pydantic[email]==2.3.0
python-jose==3.5.0